+++ inmobiliaria/src/modelo/Inmueble.java
@@ -6,12 +6,12 @@ import java.util.Locale;
 public class Inmueble {
     /* Atributos */
     private int id;
-    private int precio;
+    private long precio;
     private String direccion;
     private double mts2;
 
     /* Constructor */
-    public Inmueble(int id, int precio, String direccion, double mts2) {
+    public Inmueble(int id, long precio, String direccion, double mts2) {
         this.id = id;
         this.precio = precio;
         this.direccion = direccion;
@@ -35,11 +35,11 @@ public class Inmueble {
         this.id = id;
     }
 
-    public double getPrecio() {
+    public long getPrecio() {
         return precio;
     }
 
-    public void setPrecio(int precio) {
+    public void setPrecio(long precio) {
         this.precio = precio;
     }
 
@@ -65,8 +65,8 @@ public class Inmueble {
 
     /* Metodos */
 
-    public int precioMetroCuadrado() {
-        return (int) this.getMts2() * 110282;  /*Valor mt2 2024 */
+    public long precioMetroCuadrado() {
+        return (long) this.getMts2() * 110282;  /*Valor mt2 2024 */
     }
 
     public String getPrecioFormat() {
